frontend/api/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langgraph>=0.1.0
langchain-cerebras>=0.1.0
httpx>=0.24.0
